# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IndiaConnect Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/buyer:
    post:
      operationId: registerBuyer
      tags: [auth]
      summary: Register a new buyer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerRegistrationInput"
      responses:
        "201":
          description: Buyer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email or phone already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/seller:
    post:
      operationId: registerSeller
      tags: [auth]
      summary: Register a seller with Aadhaar and GST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerRegistrationInput"
      responses:
        "201":
          description: Seller registered, pending admin approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid Aadhaar/GST format or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email/phone/GST already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP for phone login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneLoginInput"
      responses:
        "200":
          description: OTP sent (simulated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerificationInput"
      responses:
        "200":
          description: OTP verified, logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithSeller"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Categories ────────────────────────────────────────────────────────
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Products ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    BuyerRegistrationInput:
      type: object
      required: [name, auth_method]
      properties:
        name:
          type: string
          minLength: 2
        auth_method:
          type: string
          enum: [email, phone]
        email:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    SellerRegistrationInput:
      type: object
      required: [name, email, password, business_name, aadhaar_number, gst_number, city, state, pincode]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 8
        phone:
          type: ["string", "null"]
        business_name:
          type: string
          minLength: 2
        business_description:
          type: ["string", "null"]
        aadhaar_number:
          type: string
          description: 12-digit Aadhaar number
        gst_number:
          type: string
          description: 15-character GST number
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    PhoneLoginInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpVerificationInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserWithSeller"
        message:
          type: string
      required:
        - user
    OtpRequestResponse:
      type: object
      properties:
        message:
          type: string
        dev_otp:
          type: ["string", "null"]
          description: Shown in dev mode only
      required:
        - message
...
    SellerProfile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        business_name:
          type: string
        business_description:
          type: ["string", "null"]
        aadhaar_last4:
          type: string
        gst_number:
          type: string
        verification_status:
          type: string
          enum: [pending, approved, rejected]
        rejection_reason:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - business_name
        - aadhaar_last4
        - gst_number
        - verification_status
        - city
        - state
        - pincode
        - created_at
    UserWithSeller:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [buyer, seller, admin]
        seller_profile:
          oneOf:
            - $ref: "#/components/schemas/SellerProfile"
            - type: "null"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - created_at

    # ── Category schemas ────────────────────────────────────────────────
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        product_count:
          type: integer
        service_count:
          type: integer
      required:
        - id
        - name
        - slug
        - icon
        - product_count
        - service_count

    # ── Product schemas ─────────────────────────────────────────────────
...