# lib/api-zod/src/generated/types/userWithSeller.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * IndiaConnect Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { SellerProfile } from './sellerProfile';
import type { UserWithSellerRole } from './userWithSellerRole';

export interface UserWithSeller {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: UserWithSellerRole;
  seller_profile?: SellerProfile | null;
  created_at: Date;
}